Code for Tilt Ball switch project
This is the Arduino Code for a little project that utilizes a tilt ball switch. When the the board that the project is made on tilts, it allows you to press a button that turns on an LED light. There is another button that turns the LED light off. You only press the button when the board is tilted otherwise, the LED wont turn on. So basically a security measure.